moved tests to better fitting namespace
