+++ Tests.LightBlueFox.Connect/Serialization/CompositeSerializers.cs
@@ -9,7 +9,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Tests.LightBlueFox.Connect.CustomProtocol
+namespace Tests.LightBlueFox.Connect.Serialization
 {
     [TestClass]
     public class CompositeSerializers
